refactor(rule): dry
